refactor(widgets): add correct changes as recommended
rename files/folders
