+++ src/widgets/join-us/index.ts
@@ -1,0 +1,1 @@
+export { JoinUs } from './ui/join-us'
